created UserNotFoundException as well as several functions in UserController
updated getAllUsers routing from "/user" to "/users"

+++ fullstack-backend/src/main/java/com/restaurant/fullstack_backend/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.restaurant.fullstack_backend.controller;
 
+import com.restaurant.fullstack_backend.Exceptions.UserNotFoundException;
 import com.restaurant.fullstack_backend.model.User;
 import com.restaurant.fullstack_backend.repository.UserRepository;
 
@@ -11,6 +12,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -27,14 +29,36 @@ public class UserController {
         return userRepository.save(newUser);
     }
 
-    @GetMapping("/user")
+    @GetMapping("/users")
     List<User> getAllUsers()
     {
         return userRepository.findAll();
     }
 
-        @DeleteMapping("/user/{id}")
+    @GetMapping("/user/{id}")
+    User getUserById(@PathVariable Long id){
+        return userRepository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
+    }
+
+    @PutMapping("/user/{id}")
+    User updateUser(@RequestBody User newUser, @PathVariable Long id){
+        return userRepository.findById(id).map(user -> {
+        user.setReservationId(newUser.getReservationId());
+        user.setFirstName(newUser.getFirstName());
+        user.setguestAmount(newUser.getguestAmount());
+        user.setPhoneNumber(newUser.getPhoneNumber());
+        user.setcustomerEmail(newUser.getcustomerEmail());
+        user.setreservationDate(newUser.getreservationDate());
+        user.settableNumber(newUser.gettableNumber());
+        return userRepository.save(user);}).orElseThrow(() -> new UserNotFoundException(id));
+}
+
+    @DeleteMapping("/user/{id}")
     String deleteUser(@PathVariable Long id){
+
+        if(!userRepository.existsById(id)){
+            throw new UserNotFoundException(id);
+        }
         userRepository.deleteById(id);
         return  "Group #"+id+" have been deleted.";
     }
